[Refactor] Move ResourceManager to common
