+++ common/setup.py
@@ -21,6 +21,7 @@ install_requires = [
     'numpy',
     'pandas',
     'boto3',
+    'psutil',
     'setuptools',
 ]
