+++ common/src/autogluon/common/utils/utils.py
@@ -124,3 +124,9 @@ def compare_autogluon_metadata(*, original: dict, current: dict, check_packages=
         logger.log(log[0], f"\t{log[1]}")
 
     return logs
+
+
+def bytes_to_mega_bytes(memory_amount: int) -> int:
+    """ Utility to convert a number of bytes (int) into a number of mega bytes (int)
+    """
+    return memory_amount >> 20
