+++ core/src/autogluon/core/utils/utils.py
@@ -1,7 +1,4 @@
 import logging
-import multiprocessing
-import os
-import subprocess
 import math
 import pickle
 import time
@@ -17,6 +14,7 @@ from pandas import DataFrame, Series
 from sklearn.model_selection import RepeatedKFold, RepeatedStratifiedKFold, LeaveOneGroupOut
 from sklearn.model_selection import train_test_split
 
+from autogluon.common.utils.resource_utils import ResourceManager
 from .miscs import warning_filter
 from ..constants import (
     BINARY, LARGE_DATA_THRESHOLD, MULTICLASS, MULTICLASS_UPPER_LIMIT, QUANTILE,
@@ -24,95 +22,7 @@ from ..constants import (
 )
 from ..metrics import accuracy, root_mean_squared_error, pinball_loss, Scorer
 
-
 logger = logging.getLogger(__name__)
-
-
-class ResourceManager():
-    """Manager that fetches system related info"""
-    
-    @staticmethod
-    def get_cpu_count():
-        return multiprocessing.cpu_count()
-    
-    @staticmethod
-    def get_cpu_count_psutil(logical=True):
-        return psutil.cpu_count(logical=logical)
-    
-    @staticmethod
-    def get_gpu_count_all():
-        num_gpus = ResourceManager._get_gpu_count_cuda()
-        if num_gpus == 0:
-            # Get num gpus from mxnet first because of https://github.com/autogluon/autogluon/issues/2042
-            # TODO: stop using mxnet to determine num gpus once mxnet is removed from AG
-            num_gpus = ResourceManager.get_gpu_count_mxnet()
-            if num_gpus == 0:
-                num_gpus = ResourceManager.get_gpu_count_torch()
-        return num_gpus
-
-    @staticmethod
-    def get_gpu_count_mxnet():
-        # TODO: Remove this once AG get rid off mxnet
-        try:
-            import mxnet
-            num_gpus = mxnet.context.num_gpus()
-        except Exception:
-            num_gpus = 0
-        return num_gpus
-    
-    @staticmethod
-    def get_gpu_count_torch():
-        try:
-            import torch
-            num_gpus = torch.cuda.device_count()
-        except Exception:
-            num_gpus = 0
-        return num_gpus
-    
-    @staticmethod
-    def get_gpu_free_memory():
-        """Grep gpu free memory from nvidia-smi tool.
-        This function can fail due to many reasons(driver, nvidia-smi tool, envs, etc) so please simply use
-        it as a suggestion, stay away with any rules bound to it.
-        E.g. for a 4-gpu machine, the result can be list of int
-        >>> print(get_gpu_free_memory)
-        >>> [13861, 13859, 13859, 13863]
-        """
-        _output_to_list = lambda x: x.decode('ascii').split('\n')[:-1]
-
-        try:
-            COMMAND = "nvidia-smi --query-gpu=memory.free --format=csv"
-            memory_free_info = _output_to_list(subprocess.check_output(COMMAND.split()))[1:]
-            memory_free_values = [int(x.split()[0]) for i, x in enumerate(memory_free_info)]
-        except:
-            memory_free_values = []
-        return memory_free_values
-    
-    @staticmethod
-    def get_memory_size():
-        return bytes_to_mega_bytes(psutil.virtual_memory().total)
-    
-    @staticmethod
-    def get_available_disk_size():
-        # FIXME: os.statvfs doesn't work on Windows... 
-        # Need to find another way to calculate disk on Windows.
-        # Return None for now
-        try:
-            statvfs = os.statvfs(".")
-            available_blocks = statvfs.f_frsize * statvfs.f_bavail
-            return bytes_to_mega_bytes(available_blocks)
-        except Exception:
-            return None
-    
-    @staticmethod
-    def _get_gpu_count_cuda():
-        # FIXME: Sometimes doesn't detect GPU on Windows
-        # FIXME: Doesn't ensure the GPUs are actually usable by the model (MXNet, PyTorch, etc.)
-        from .nvutil import cudaInit, cudaDeviceGetCount, cudaShutdown
-        if not cudaInit(): return 0
-        gpu_count = cudaDeviceGetCount()
-        cudaShutdown()
-        return gpu_count
 
 
 class CVSplitter:
@@ -1022,9 +932,3 @@ def unevaluated_fi_df_template(features: List[str]) -> pd.DataFrame:
         'n': 0
     }, index=features)
     return importance_df
-
-
-def bytes_to_mega_bytes(memory_amount: int) -> int:
-    """ Utility to convert a number of bytes (int) into a number of mega bytes (int)
-    """
-    return memory_amount >> 20
