+++ core/tests/unittests/hpo/test_ray_hpo.py
@@ -9,7 +9,6 @@ from autogluon.core.hpo.ray_hpo import (
     AutommRayTuneLightningAdapter,
     run,
 )
-from autogluon.core.utils import ResourceManager
 from autogluon.core.hpo.ray_tune_constants import SEARCHER_PRESETS, SCHEDULER_PRESETS
 from ray import tune
