+++ core/tests/unittests/test_parallel_local_folding.py
@@ -5,7 +5,7 @@ import pandas as pd
 
 from autogluon.core.models.ensemble.bagged_ensemble_model import BaggedEnsembleModel
 from autogluon.core.models.ensemble.fold_fitting_strategy import ParallelLocalFoldFittingStrategy
-from autogluon.core.utils import ResourceManager
+from autogluon.common.utils.resource_utils import ResourceManager
 from autogluon.core.models import AbstractModel
 from autogluon.core.searcher import LocalRandomSearcher
 from autogluon.core.space import Int
