+++ features/setup.py
@@ -21,7 +21,6 @@ install_requires = [
     'numpy',
     'pandas',
     'scikit-learn',
-    'psutil',
     f'autogluon.common=={version}',
 ]
