+++ tabular/src/autogluon/tabular/models/text_prediction/text_prediction_v1_model.py
@@ -7,8 +7,8 @@ import pandas as pd
 
 from autogluon.common.features.types import R_OBJECT, R_INT, R_FLOAT, R_CATEGORY, \
     S_TEXT_NGRAM, S_TEXT_AS_CATEGORY, S_TEXT_SPECIAL, S_IMAGE_PATH
+from autogluon.common.utils.resource_utils import ResourceManager
 from autogluon.core.constants import REGRESSION
-from autogluon.core.utils import ResourceManager
 from autogluon.core.utils import try_import_autogluon_text
 from autogluon.core.models import AbstractModel
