+++ tabular/src/autogluon/tabular/models/xgboost/xgboost_model.py
@@ -5,10 +5,10 @@ import psutil
 
 from autogluon.common.features.types import R_BOOL, R_INT, R_FLOAT, R_CATEGORY
 from autogluon.common.utils.pandas_utils import get_approximate_df_mem_usage
+from autogluon.common.utils.resource_utils import ResourceManager
 from autogluon.core.constants import MULTICLASS, REGRESSION, SOFTCLASS, PROBLEM_TYPES_CLASSIFICATION
 from autogluon.core.models import AbstractModel
 from autogluon.core.models._utils import get_early_stopping_rounds
-from autogluon.core.utils import ResourceManager
 from autogluon.core.utils import try_import_xgboost
 from autogluon.core.utils.exceptions import NotEnoughMemoryError
