+++ tabular/tests/conftest.py
@@ -7,7 +7,7 @@ import pytest
 from contextlib import contextmanager
 from typing import List
 
-from autogluon.core.utils import ResourceManager
+from autogluon.common.utils.resource_utils import ResourceManager
 from autogluon.core.utils import download, unzip
 from autogluon.core.constants import BINARY, MULTICLASS, REGRESSION
 from autogluon.core.data.label_cleaner import LabelCleaner
