+++ tabular/tests/unittests/resource_allocation/test_total_resource_allocation.py
@@ -1,6 +1,6 @@
 import pytest
 
-from autogluon.core.utils import ResourceManager
+from autogluon.common.utils.resource_utils import ResourceManager
 from autogluon.core.models.ensemble.bagged_ensemble_model import BaggedEnsembleModel
 from autogluon.tabular.models import AbstractModel
